Add a confirmation prompt to database.reset

+++ app/cli/database.py
@@ -9,9 +9,16 @@ def database():
 
 
 @database.command('reset', help='Reset everything')
-def reset():
+@click.option(
+    '-y', '--confirm', 'confirm',
+    is_flag=True, default=False,
+    help='Skip the prompt asking to confirm the command'
+)
+def reset(confirm: bool):
+    if not confirm:
+        click.confirm(click.style('Do you really want to reset the whole database?', fg='bright_red'), abort=True)
     try:
         db.drop_all()
-        click.echo('The database has been reset!')
+        click.secho('The database has been reset!', fg='bright_green')
     except Exception as e:
-        click.echo(f'Error {e.__class__.__name__}: {e}')
+        click.secho(f'Error {e.__class__.__name__}: {e}', fg='bright_red')
